removed some more warnings

+++ PndTrackers/PndForwardTrackFinder/PndFtsExpandedTrackCand.cxx
@@ -13,9 +13,9 @@ PndFtsExpandedTrackCand::~PndFtsExpandedTrackCand() {
 
 PndFtsExpandedTrackCand::PndFtsExpandedTrackCand(vector<PndFtsHit*> h1, vector<PndFtsHit*> h2, vector<PndLineApproximation> l){
 	vector<PndFtsHit*> hits;
-	for(int i=0;i<h1.size();i++) hits.push_back(h1[i]);
-	for(int i=0;i<h2.size();i++) hits.push_back(h2[i]);
-	for(int i=0;i<hits.size();i++){
+	for(size_t i=0;i<h1.size();i++) hits.push_back(h1[i]);
+	for(size_t i=0;i<h2.size();i++) hits.push_back(h2[i]);
+	for(size_t i=0;i<hits.size();i++){
 		fPndTrackCand.AddHit(hits[i]->GetEntryNr(),i);
 	}
 	fSourceHits = hits;
